Slight refactoring of simplify
